+++ src/ncel/Log.cs
@@ -6,5 +6,9 @@
         {
             Recorder.StoreLineInFile(LogLevel.Information, msgToLog);
         }
+        public static void Error(string msgToLog)
+        {
+            Recorder.StoreLineInFile(LogLevel.Error, msgToLog);
+        }
     }
 }