+++ src/ncel/LogConfig.cs
@@ -1,0 +1,7 @@
+namespace Ncel
+{
+    public static class LogConfig
+    {
+        public static bool WriteConsole { get; set; } = false;
+    }
+}