+++ test/unitTest/LogTest.cs
@@ -1,6 +1,7 @@
 using System;
 using System.IO;
 using System.Linq;
+using Bogus;
 using Ncel;
 using Xunit;
 
@@ -8,13 +9,15 @@ namespace unitTest
 {
     public class LogTest
     {
+        Faker faker = new Faker("pt_BR");
         [Fact]
         public void LogGeneratesFile()
         {
             //Given
             var x = new FileInfo(Utilities.DestinationPath());
+            x.Delete();
             //When
-            Log.Information("123 Testando");
+            Log.Information($"{faker.Random.Number(1,100)} Testando");
             //Then
             Assert.True(x.Exists);
         }
@@ -22,12 +25,37 @@ namespace unitTest
         public void LogRecordProperInformationInfile()
         {
             //Given
-            var x = File.ReadLines(Utilities.DestinationPath()).Last();
-            var y = "123 Testando";
+            var x = $"{faker.Random.Number(1,100)} Testando";
             //When
-            Log.Information(y);
+            Log.Information(x);
             //Then
-            Assert.True(x.Contains(y));
+            Assert.Contains(x, File.ReadLines(Utilities.DestinationPath()).Last());
+        }
+        [Fact]
+        public void LogConsoleMessageActivate()
+        {
+            //Given
+            var x = $"{faker.Random.Number(1,100)} Testando";
+            StringWriter y = new StringWriter();
+            Console.SetOut(y);
+            //When
+            LogConfig.WriteConsole = true;
+            Log.Information(x);
+            //Then
+            Assert.Contains(x, y.ToString());
+        }
+        [Fact]
+        public void LogConsoleMessageDeactivated()
+        {
+            //Given
+            var x = $"{faker.Random.Number(1,100)} Testando";
+            StringWriter y = new StringWriter();
+            Console.SetOut(y);
+            //When
+            LogConfig.WriteConsole = false;
+            Log.Information(x);
+            //Then
+            Assert.DoesNotContain(x, y.ToString());
         }
     }
 }